Factor out a separate library of static C++ code into sub-project 'pgen-lib'.
