+++ pgen-lib/src/main/cpp/pgenlib_misc.cc
@@ -15,7 +15,7 @@
 // along with this library.  If not, see <http://www.gnu.org/licenses/>.
 
 
-#include "include/pgenlib_misc.h"
+#include "pgenlib_misc.h"
 
 #ifdef __cplusplus
 namespace plink2 {
