+++ pgen-lib/src/main/cpp/pgenlib_read.cc
@@ -15,7 +15,7 @@
 // along with this library.  If not, see <http://www.gnu.org/licenses/>.
 
 
-#include "include/pgenlib_read.h"
+#include "pgenlib_read.h"
 
 #include <errno.h>
