+++ pgen-lib/src/main/headers/pgenlib_ffi_support.h
@@ -17,7 +17,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this library.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "include/pgenlib_misc.h"
+#include "pgenlib_misc.h"
 
 #ifdef __cplusplus
 namespace plink2 {
